Move postNewUser to userSlice
Moving code to where I think it belongs
