+++ app/src/thunks/userThunks.ts
@@ -1,5 +1,4 @@
 import axios from "axios";
-import { buildUserBody } from "../utils/resourceUtils";
 import { transformToIdValueArray } from "../utils/utils";
 import { addNotification } from "./notificationThunks";
 
@@ -12,29 +11,6 @@ export const fetchUsersAndUsernames = async () => {
 	const response = await data.data;
 
 	return transformToIdValueArray(response);
-};
-
-// new user to backend
-// @ts-expect-error TS(7006): Parameter 'values' implicitly has an 'any' type.
-export const postNewUser = (values) => async (dispatch) => {
-	// get URL params used for post request
-	let data = buildUserBody(values);
-
-	// POST request
-	axios
-		.post("/admin-ng/users", data, {
-			headers: {
-				"Content-Type": "application/x-www-form-urlencoded",
-			},
-		})
-		.then((response) => {
-			console.info(response);
-			dispatch(addNotification("success", "USER_ADDED"));
-		})
-		.catch((response) => {
-			console.error(response);
-			dispatch(addNotification("error", "USER_NOT_SAVED"));
-		});
 };
 
 // delete user with provided id
